dom/indexedDB/ipc/test_ipc.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for OOP IndexedDB</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
  <body>

  <script type="application/javascript;version=1.7">
    "use strict";

    SimpleTest.waitForExplicitFinish();

    function iframeScriptFirst() {
      content.wrappedJSObject.RunSet.reloadAndRunAll({
        preventDefault: function() { }
      });
    }

    function iframeScriptSecond() {
      let TestRunner = content.wrappedJSObject.TestRunner;

      let oldComplete = TestRunner.onComplete;

      TestRunner.onComplete = function() {
        TestRunner.onComplete = oldComplete;

        sendAsyncMessage("test:indexedDB:ipcTestComplete", {
          result: JSON.stringify(TestRunner._failedTests)
        });

        if (oldComplete) {
          oldComplete();
        }
      };

      let oldLog = TestRunner.log;
      TestRunner.log = function(msg) {
        sendAsyncMessage("test:indexedDB:ipcTestMessage", { msg: msg });
      }
    }

    let regex = /^(TEST-PASS|TEST-UNEXPECTED-PASS|TEST-KNOWN-FAIL|TEST-UNEXPECTED-FAIL|TEST-DEBUG-INFO) \| ([^\|]+) \|(.*)/;

    function onTestMessage(data) {
      let message = data.json.msg;
      let match = regex.exec(message);
      if (match) {
        let state = match[1];
        let details = match[2] + " | " + match[3];

        switch (state) {
          case "TEST-PASS":
          case "TEST-KNOWN-FAIL":
            ok(true, details);
            break;

          case "TEST-UNEXPECTED-FAIL":
          case "TEST-UNEXPECTED-PASS":
            ok(false, details);
            break;

          case "TEST-DEBUG-INFO":
          default:
            info(details);
        }
      }
    }

    function onTestComplete() {
      let comp = SpecialPowers.wrap(Components);

      let idbManager =
        comp.classes["@mozilla.org/dom/indexeddb/manager;1"]
            .getService(comp.interfaces.nsIIndexedDatabaseManager);

      let uri = SpecialPowers.getDocumentURIObject(document);
      idbManager.clearDatabasesForURI(uri);

      idbManager.getUsageForURI(uri, function(uri, usage, fileUsage) {
        is(usage, 0, "Cleared ipc databases properly");
        SimpleTest.executeSoon(function () { SimpleTest.finish(); });
      });
    }

    function runTests() {
      let iframe = document.createElement("iframe");
      iframe.mozbrowser = true;
      iframe.id = "iframe";
      iframe.style.width = "100%";
      iframe.style.height = "1000px";

      function iframeLoadSecond() {
        ok(true, "Got second iframe load event.");
        iframe.removeEventListener("mozbrowserloadend", iframeLoadSecond);
        let mm = SpecialPowers.getBrowserFrameMessageManager(iframe);
        mm.loadFrameScript("data:,(" + iframeScriptSecond.toString() + ")();",
                           false);
      }

      function iframeLoadFirst() {
        ok(true, "Got first iframe load event.");
        iframe.removeEventListener("mozbrowserloadend", iframeLoadFirst);
        iframe.addEventListener("mozbrowserloadend", iframeLoadSecond);

        let mm = SpecialPowers.getBrowserFrameMessageManager(iframe);

        let comp = SpecialPowers.wrap(Components);

        let spObserver =
          comp.classes["@mozilla.org/special-powers-observer;1"]
              .getService(comp.interfaces.nsIFrameMessageListener);

        mm.addMessageListener("SPPrefService", spObserver);
        mm.addMessageListener("SPProcessCrashService", spObserver);
        mm.addMessageListener("SPPingService", spObserver);
        mm.addMessageListener("SpecialPowers.Quit", spObserver);
        mm.addMessageListener("SPPermissionManager", spObserver);

        mm.addMessageListener("test:indexedDB:ipcTestMessage", onTestMessage);
        mm.addMessageListener("test:indexedDB:ipcTestComplete", onTestComplete);

        let specialPowersBase = "chrome://specialpowers/content/";
        mm.loadFrameScript(specialPowersBase + "MozillaLogger.js", false);
        mm.loadFrameScript(specialPowersBase + "specialpowersAPI.js", false);
        mm.loadFrameScript(specialPowersBase + "specialpowers.js", false);

        mm.loadFrameScript("data:,(" + iframeScriptFirst.toString() + ")();",
                           false);
      }

      iframe.addEventListener("mozbrowserloadend", iframeLoadFirst);

      // Strip this filename and one directory level and then add "/test".
      let href =  window.location.href;
      href = href.substring(0, href.lastIndexOf('/'));
      href = href.substring(0, href.lastIndexOf('/'));
      iframe.src = href + "/test?consoleLevel=INFO";

      document.body.appendChild(iframe);
    }

    addEventListener("load", function() {
      let whitelist;
      try {
        whitelist =
          SpecialPowers.getCharPref("dom.mozBrowserFramesWhitelist") + ", ";
      } catch (e) {
        whitelist = "";
      }

      whitelist += window.location.protocol + "//" + window.location.host;

      SpecialPowers.pushPrefEnv({
        "set": [
          ["dom.ipc.browser_frames.oop_by_default", true],
          ["dom.mozBrowserFramesEnabled", true],
          ["dom.mozBrowserFramesWhitelist", whitelist],
          ["browser.pageThumbs.enabled", false]
        ]
      }, runTests);
    });
  </script>
</body>
</html>
